+++ packages/searchbar/src/Searchbar.tsx
@@ -8,7 +8,7 @@ import { styled } from '@pikas-ui/styles'
 import type { TextfieldProps, TextfieldStylesType } from '@pikas-ui/textfield'
 import { Textfield } from '@pikas-ui/textfield'
 import React, { useEffect, useState, useRef } from 'react'
-import usehook from 'usehooks-ts'
+import * as usehook from 'usehooks-ts'
 
 const { useDebounce, useOnClickOutside, useWindowSize } = usehook