+++ packages/styles/src/PikasUIProvider/PikasUIProvider.tsx
@@ -3,7 +3,7 @@ import type { createTheme } from '../css.js'
 import { styled, theme as themeDefault } from '../css.js'
 import merge from 'lodash.merge'
 import cloneDeep from 'lodash.clonedeep'
-import usehooks from 'usehooks-ts'
+import * as usehooks from 'usehooks-ts'
 
 const { useTernaryDarkMode, useSsr } = usehooks