+++ packages/table/src/pagination/Pagination.tsx
@@ -114,11 +114,7 @@ export const Pagination: React.FC<PaginationProps> = ({
                 backgroundColor: '$PRIMARY',
                 color:
                   (theme &&
-                    // @ts-ignore
-                    fontColorContrast.default(
-                      theme.colors['PRIMARY'].value,
-                      0.7
-                    )) ||
+                    fontColorContrast(theme.colors['PRIMARY'].value, 0.7)) ||
                   undefined,
               }),
             }}