+++ packages/table/src/tfoot/Tfoot.tsx
@@ -19,8 +19,7 @@ export const Tfoot: React.FC<CustomProps> = (props) => {
           backgroundColor: '$PRIMARY',
           color:
             (theme &&
-              // @ts-ignore
-              fontColorContrast.default(theme.colors['PRIMARY'].value, 0.7)) ||
+              fontColorContrast(theme.colors['PRIMARY'].value, 0.7)) ||
             undefined,
 
           tr: {