+++ packages/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/dropdown-menu
 
+## 2.1.11
+
+### Patch Changes
+
+- Fix ESM
+- Updated dependencies
+  - @pikas-ui/icons@2.1.10
+  - @pikas-ui/loader@2.1.10
+  - @pikas-ui/menu@2.1.10
+  - @pikas-ui/styles@2.3.10
+
 ## 2.1.10
 
 ### Patch Changes