+++ packages/searchbar/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pikas-ui/searchbar
 
+## 2.1.13
+
+### Patch Changes
+
+- Fix ESM
+- Updated dependencies
+  - @pikas-ui/button@2.1.10
+  - @pikas-ui/icons@2.1.10
+  - @pikas-ui/loader@2.1.10
+  - @pikas-ui/separator@2.1.10
+  - @pikas-ui/styles@2.3.10
+  - @pikas-ui/textfield@2.1.13
+
 ## 2.1.12
 
 ### Patch Changes