+++ packages/select/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/select
 
+## 3.1.13
+
+### Patch Changes
+
+- Fix ESM
+- Updated dependencies
+  - @pikas-ui/icons@2.1.10
+  - @pikas-ui/styles@2.3.10
+  - @pikas-ui/text@3.1.10
+  - @pikas-ui/textfield@2.1.13
+  - @pikas-ui/tooltip@2.1.10
+
 ## 3.1.12
 
 ### Patch Changes