+++ packages/slider/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/slider
 
+## 2.1.10
+
+### Patch Changes
+
+- Fix ESM
+- Updated dependencies
+  - @pikas-ui/icons@2.1.10
+  - @pikas-ui/styles@2.3.10
+  - @pikas-ui/text@3.1.10
+
 ## 2.1.9
 
 ### Patch Changes