+++ packages/switch/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/switch
 
+## 1.1.11
+
+### Patch Changes
+
+- Fix ESM
+- Updated dependencies
+  - @pikas-ui/icons@2.1.10
+  - @pikas-ui/styles@2.3.10
+  - @pikas-ui/text@3.1.10
+  - @pikas-ui/tooltip@2.1.10
+
 ## 1.1.10
 
 ### Patch Changes