+++ packages/table/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pikas-ui/table
 
+## 2.1.13
+
+### Patch Changes
+
+- Fix ESM
+- Updated dependencies
+  - @pikas-ui/checkbox@2.1.11
+  - @pikas-ui/icons@2.1.10
+  - @pikas-ui/loader@2.1.10
+  - @pikas-ui/select@3.1.13
+  - @pikas-ui/styles@2.3.10
+
 ## 2.1.12
 
 ### Patch Changes