+++ packages/tabs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/tabs
 
+## 1.0.8
+
+### Patch Changes
+
+- Fix ESM
+- Updated dependencies
+  - @pikas-ui/styles@2.3.10
+
 ## 1.0.7
 
 ### Patch Changes